🕷️  fix Vagrantfile
It seems the ruby yaml parser doesn't parse yaml aliases unless it gets
explicitly configured to do so. A change to .gitlab-ci.yml has
introduced a yaml alias and that in turn broke the `Vagrantfile`.

Without this change I was getting this error message:

```
% vagrant status
Vagrant failed to initialize at a very early stage:

There was an error loading a Vagrantfile. The file being loaded
and the error message are shown below. This is usually caused by
a syntax error.

Path: .../tor-android/Vagrantfile
Line number: 0
Message: Psych::BadAlias: Unknown alias: job-template
```